+++ vm/asyncio.go
@@ -131,6 +131,8 @@ func (i *Interp) builtinModule(name string) (*object.Module, bool) {
 		return i.buildPoplib(), true
 	case "imaplib":
 		return i.buildImaplib(), true
+	case "smtplib":
+		return i.buildSmtplib(), true
 	case "http":
 		return i.buildHttp(), true
 	case "http.client":